Add WIP cursor
This is a wip cursor made by me, it's lazer styled but blue. Hubz is currently working on a better version.
If you wanna try it in-game now, download it and put it in your skin folder as "cursor.png". You might also need to delete cursortrail@2x.png if you use an HD screen.